Don't spread props onto DOM elements, fixes #35
Currently all props are being spread onto the wrapping `div` elements, this is causing react to complain about unknown props on DOM elements. I can't see an obvious reason for the spread given that props intended for the lazy-loaded component should not need to be present on a wrapper element. So, this PR removes the spreads entirely.